+++ src/block.js
@@ -18,6 +18,8 @@ var EventBus = require('./event-bus');
 
 var { Spinner } = require('spin.js');
 
+var { trimScribeContent } = require('./blocks/scribe-plugins/shared');;
+
 const DELETE_TEMPLATE = require("./templates/delete");
 const BLOCK_DESCRIPTION_TEMPLATE = require("./templates/block-description");
 
@@ -228,6 +230,10 @@ Object.assign(Block.prototype, SimpleBlock.fn, require('./block-validations'), {
     });
   },
 
+  focusAtStart: function() {
+    this.focus();
+  },
+
   focusAtEnd: function() {
     this.focus();
   },
@@ -256,10 +262,6 @@ Object.assign(Block.prototype, SimpleBlock.fn, require('./block-validations'), {
   },
 
   _onBlur: function() {},
-
-  onBlockRender: function() {
-    this.focus();
-  },
 
   onDrop: function(dataTransferObj) {},
 
@@ -446,29 +448,22 @@ Object.assign(Block.prototype, SimpleBlock.fn, require('./block-validations'), {
   setTextBlockHTML: function(html) {
     var returnVal = this._scribe.setContent(html);
 
-    // Remove any whitespace in the first node, otherwise selections won't work.
-    var firstNode = this._scribe.node.firstDeepestChild(this._scribe.el);
-    if (firstNode.nodeName === '#text') {
-      firstNode.textContent = utils.leftTrim(firstNode.textContent);
-    }
-
-    // Remove all empty nodes at the front to get blocks working.
-    // Don't remove nodes that can't contain text content (e.g. <input>)
-    while (this._scribe.el.firstChild && this._scribe.el.firstChild.textContent === '' && document.createElement(this._scribe.el.firstChild.tagName).outerHTML.indexOf("/") != -1) {
-      this._scribe.el.removeChild(this._scribe.el.firstChild);
-    }
-
-    // Firefox adds empty br tags at the end of content.
-    while(this._scribe.el.lastChild && this._scribe.el.lastChild.nodeName === 'BR') {
-      this._scribe.el.removeChild(this._scribe.el.lastChild);
-    }
+    trimScribeContent(this._scribe);
 
     return returnVal;
   },
 
   isEmpty: function() {
     return _.isEmpty(this.getBlockData());
-  }
+  },
+
+  select: function(selected) {
+    this.el.classList.toggle("st-block--is-selected", selected);
+  },
+
+  split: function() {},
+
+  asClipboardHTML: function() {}
 
 });