+++ src/blocks/scribe-plugins/scribe-list-block-plugin.js
@@ -6,6 +6,9 @@ var {
   getTotalLength,
   isAtStart,
   isAtEnd,
+  isSelectedFromStart,
+  isSelectedToEnd,
+  rangeToHTML,
   selectToEnd
 } = require('./shared.js');
 
@@ -17,60 +20,46 @@ var ScribeListBlockPlugin = function(block) {
         return;
       }
 
-      var rangeToHTML = function(range) {
-        var div = document.createElement('div');
-        div.appendChild(range.extractContents());
-
-        return div.innerHTML;
-      };
-
       var content;
 
       if (ev.key === "Enter" && !ev.shiftKey) {
         ev.preventDefault();
 
-        if (scribe.getTextContent().length === 0) {
-          let nextListItem = block.nextListItem();
-          if (nextListItem) {
-            const data = {format: 'html', listItems: []};
-            block.removeCurrentListItem();
-            block.focusOn(nextListItem);
-            while (!!nextListItem) {
-              data.listItems.push({content: nextListItem.scribe.getContent()});
-              block.focusOn(nextListItem);
-              block.removeCurrentListItem();
-              nextListItem = block.nextListItem();
-            }
-            block.mediator.trigger("block:create", 'List', data, block.el, { autoFocus: true });
-            block.mediator.trigger("block:create", 'Text', null, block.el, { autoFocus: true });
+        block.splitListItem(scribe, { createTextBlock: true });
+      } else if (["Left", "ArrowLeft", "Up", "ArrowUp"].indexOf(ev.key) > -1) {
+        if (ev.shiftKey && isSelectedFromStart(scribe)) {
+          ev.preventDefault();
+          ev.stopPropagation();
+
+          document.activeElement && document.activeElement.blur();
+          block.mediator.trigger("selection:block", block);
+        } else if (isAtStart(scribe)) {
+          ev.preventDefault();
+
+          var previousListItem = block.previousListItem();
+          if (previousListItem) {
+            block.focusOn(previousListItem, { focusAtEnd: true });
           } else {
-            block.removeCurrentListItem();
-            block.mediator.trigger("block:create", 'Text', null, block.el, { autoFocus: true });
+            block.mediator.trigger("block:focusPrevious", block.blockID);
           }
-        } else {
-          content = rangeToHTML(selectToEnd(scribe));
-          block.addListItemAfterCurrent(content);
         }
-      } else if (["Left", "ArrowLeft", "Up", "ArrowUp"].indexOf(ev.key) > -1  && isAtStart(scribe)) {
-        ev.preventDefault();
+      } else if (["Right", "ArrowRight", "Down", "ArrowDown"].indexOf(ev.key) > -1) {
+        if (ev.shiftKey && isSelectedToEnd(scribe)) {
+          ev.preventDefault();
+          ev.stopPropagation();
 
-        var previousListItem = block.previousListItem();
-        if (previousListItem) {
-          block.focusOn(previousListItem, { focusAtEnd: true });
-        } else {
-          block.mediator.trigger("block:focusPrevious", block.blockID);
+          document.activeElement && document.activeElement.blur();
+          block.mediator.trigger("selection:block", block);
+        } else if (isAtEnd(scribe)) {
+          ev.preventDefault();
+
+          var nextListItem = block.nextListItem();
+          if (nextListItem) {
+            block.focusOn(nextListItem);
+          } else {
+            block.mediator.trigger("block:focusNext", block.blockID);
+          }
         }
-
-      } else if (["Right", "ArrowRight", "Down", "ArrowDown"].indexOf(ev.key) > -1 && isAtEnd(scribe)) {
-        ev.preventDefault();
-
-        var nextListItem = block.nextListItem();
-        if (nextListItem) {
-          block.focusOn(nextListItem);
-        } else {
-          block.mediator.trigger("block:focusNext", block.blockID);
-        }
-
       } else if (ev.key === "Backspace" && isAtStart(scribe)) {
         ev.preventDefault();