+++ CHANGELOG.md
@@ -1,11 +1,43 @@
 # Changelog
 
+**Current version: 0.8.0**
+
+## 0.8.x
+
+### ????
+- [ ] Improve edge content pasting.
+- [ ] Add undo / redo functionality.
+- [ ] Allow selection of individual list items.
+- [ ] Automated testing for copy / paste.
+
+### 0.8.2 - Small fixes
+* Improve list blocks in Chrome where empty item drops down a line.
+* Added add button to bottom of quote blocks.
+
+### 0.8.1 - Add Copy / Paste
+* Selection keyboard behaviour.
+* Selection mouse behaviour.
+* Native browser selection behaviour.
+* Enable paste of non-text content.
+* Enable cut of blocks.
+* Cross browser fixes / testing.
+* Improve Firefox content pasting.
+
+### 0.8.0
+* Add Copy all to the editor
+* Implement Selection Handler
+* Block level selection.
+* Hide new functionality behing feature flags.
+
 ## 0.7.x
 
-### Future releases
+### 0.7.5
 
-* Improve Firefox content pasting.
-* Improve edge content pasting.
+* Separate block positioner select data to improve performance. https://github.com/madebymany/sir-trevor-js/pull/408
+
+### 0.7.4
+
+* Fix `config.focusOnInit` to work with non text blocks by removing default `onBlockRender` that is no longer needed.
 
 ### 0.7.3